LayoutTests/platform/mac-wk1/media/media-source/worker/media-managedmse-worker-expected.txt


EXPECTED (ManagedMediaSource.canConstructInDedicatedWorker == 'true'), OBSERVED 'false' FAIL
RUN(video.disableRemotePlayback = true)
error message from worker: MSE not supported FAIL
END OF TEST
